net: hns3: fix length and page_offset overflow when CONFIG_ARM64_64K_PAGES
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

When enable the config item "CONFIG_ARM64_64K_PAGES", the size of PAGE_SIZE
is 65536(64K). But the  type of length and page_offset are u16, they will
overflow. So change them to u32.

+++ drivers/net/ethernet/hisilicon/hns3/hns3_enet.h
@@ -303,8 +303,8 @@ struct hns3_desc_cb {
 
 	/* priv data for the desc, e.g. skb when use with ip stack*/
 	void *priv;
-	u16 page_offset;
-	u16 reuse_flag;
+	u32 page_offset;
+	u32 reuse_flag;
 
 	u32 length;     /* length of the buffer */
